requirements.txt
pika
psycopg2-binary
requests
python-dotenv